+++ osanReactStudy/package.json
@@ -14,6 +14,7 @@
     "lucide-react": "^0.489.0",
     "react": "^19.0.0",
     "react-dom": "^19.0.0",
+    "react-router-dom": "^7.5.3",
     "tailwind-merge": "^3.2.0"
   },
   "devDependencies": {
