+++ osanReactStudy/src/pages/Attendance.jsx
@@ -1,0 +1,9 @@
+export default function Attendance() {
+    return (
+        <div>
+            <section id="header" className="container-fluid mx-auto p-4">
+                <h1 className="text-2xl font-bold">Attendance</h1>
+            </section>
+        </div>
+    )
+}
